Add `Aggregate` and remove `Aggregator` traits
The `Aggregator` trait is removed in favor of `Aggregate` - a type that
can be produced by consuming an `ExactSizeIterator` of references to an
`Option` of the same type at a given tree height.

This is an improvement on the former, since it can now easily be
reasoned about the type of data that a tree accepts, and the user no
longer has to bear the overhead of instantiating "zero items".

The resulting API is also simpler in its usage, and more clearly
understandable.